put v1 in a module
