+++ aries_vcx/src/handlers/out_of_band/sender.rs
@@ -7,7 +7,7 @@ use messages::{
             invitation::{Invitation, InvitationContent, InvitationDecorators, OobService},
             OobGoalCode,
         },
-        present_proof::PresentProof,
+        present_proof::v1::PresentProof,
     },
     msg_types::Protocol,
     AriesMessage,
