+++ aries_vcx/src/protocols/proof_presentation/prover/states/presentation_request_received.rs
@@ -4,7 +4,7 @@ use aries_vcx_core::{
     anoncreds::base_anoncreds::BaseAnonCreds, ledger::base_ledger::AnoncredsLedgerRead,
 };
 use messages::msg_fields::protocols::{
-    present_proof::{
+    present_proof::v1::{
         present::Presentation,
         request::{RequestPresentation, RequestPresentationContent},
     },
