+++ aries_vcx/src/protocols/proof_presentation/verifier/states/presentation_proposal_received.rs
@@ -1,4 +1,4 @@
-use messages::msg_fields::protocols::present_proof::{
+use messages::msg_fields::protocols::present_proof::v1::{
     propose::{PresentationPreview, ProposePresentation, ProposePresentationContent},
     request::RequestPresentation,
 };
